+++ src/main/java/insane96mcp/iguanatweaksreborn/module/misc/beaconconduit/ITRBeaconMenu.java
@@ -87,7 +87,7 @@ public class ITRBeaconMenu extends AbstractContainerMenu {
     public ItemStack quickMoveStack(Player pPlayer, int pIndex) {
         ItemStack itemstack = ItemStack.EMPTY;
         Slot slot = this.slots.get(pIndex);
-        if (slot != null && slot.hasItem()) {
+        if (slot.hasItem()) {
             ItemStack itemstack1 = slot.getItem();
             itemstack = itemstack1.copy();
             if (pIndex == 0) {