+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Upcoming
+* Some beacon effects now only unlock with more pyramid height 
 * Added `/iguanatweaks get_treasure_enchantments_book` command
 * Fixed wrong effects in beacon breaking it
 * Fixed tiredness crash